Move Kokkos::HIP out of Kokkos_HIP_Space and into Kokkos_HIP
