+++ core/src/Kokkos_HIP.hpp
@@ -54,30 +54,140 @@ static_assert(false,
 
 #if defined(KOKKOS_ENABLE_HIP)
 
-//----------------------------------------------------------------------------
-//----------------------------------------------------------------------------
-
+#include <Kokkos_Layout.hpp>
 #include <Kokkos_HIP_Space.hpp>
-#include <Kokkos_Parallel.hpp>
 
-#include <HIP/Kokkos_HIP_DeepCopy.hpp>
-#include <HIP/Kokkos_HIP_Half_Impl_Type.hpp>
-#include <HIP/Kokkos_HIP_Half_Conversion.hpp>
-#include <HIP/Kokkos_HIP_Instance.hpp>
-#include <HIP/Kokkos_HIP_MDRangePolicy.hpp>
-#include <HIP/Kokkos_HIP_Parallel_Range.hpp>
-#include <HIP/Kokkos_HIP_Parallel_MDRange.hpp>
-#include <HIP/Kokkos_HIP_Parallel_Team.hpp>
-#include <HIP/Kokkos_HIP_SharedAllocationRecord.hpp>
-#include <HIP/Kokkos_HIP_UniqueToken.hpp>
+#include <hip/hip_runtime_api.h>
 
 namespace Kokkos {
+namespace Impl {
+class HIPInternal;
+}
+/// \class HIP
+/// \brief Kokkos device for multicore processors in the host memory space.
+class HIP {
+ public:
+  //------------------------------------
+  //! \name Type declarations that all Kokkos devices must provide.
+  //@{
+
+  //! Tag this class as a kokkos execution space
+  using execution_space = HIP;
+  using memory_space    = HIPSpace;
+  using device_type     = Kokkos::Device<execution_space, memory_space>;
+
+  using array_layout = LayoutLeft;
+  using size_type    = HIPSpace::size_type;
+
+  using scratch_memory_space = ScratchMemorySpace<HIP>;
+
+  HIP();
+  HIP(hipStream_t stream, bool manage_stream = false);
+
+  //@}
+  //------------------------------------
+  //! \name Functions that all Kokkos devices must implement.
+  //@{
+
+  KOKKOS_INLINE_FUNCTION static int in_parallel() {
+#if defined(__HIP_DEVICE_COMPILE__)
+    return true;
+#else
+    return false;
+#endif
+  }
+
+  /** \brief Wait until all dispatched functors complete.
+   *
+   * The parallel_for or parallel_reduce dispatch of a functor may return
+   * asynchronously, before the functor completes. This method does not return
+   * until all dispatched functors on this device have completed.
+   */
+  static void impl_static_fence(const std::string& name);
+
+  void fence(const std::string& name =
+                 "Kokkos::HIP::fence(): Unnamed Instance Fence") const;
+
+  hipStream_t hip_stream() const;
+
+  /// \brief Print configuration information to the given output stream.
+  void print_configuration(std::ostream& os, bool verbose = false) const;
+
+  /// \brief Free any resources being consumed by the device.
+  static void impl_finalize();
+
+  /** \brief  Initialize the device.
+   *
+   */
+  int hip_device() const;
+  static hipDeviceProp_t const& hip_device_prop();
+
+  static void impl_initialize(InitializationSettings const&);
+
+  static int impl_is_initialized();
+
+  //  static size_type device_arch();
+
+  static size_type detect_device_count();
+
+  static int concurrency();
+  static const char* name();
+
+  inline Impl::HIPInternal* impl_internal_space_instance() const {
+    return m_space_instance.get();
+  }
+
+  uint32_t impl_instance_id() const noexcept;
+
+ private:
+  friend bool operator==(HIP const& lhs, HIP const& rhs) {
+    return lhs.impl_internal_space_instance() ==
+           rhs.impl_internal_space_instance();
+  }
+  friend bool operator!=(HIP const& lhs, HIP const& rhs) {
+    return !(lhs == rhs);
+  }
+  Kokkos::Impl::HostSharedPtr<Impl::HIPInternal> m_space_instance;
+};
+
+namespace Impl {
+template <>
+struct MemorySpaceAccess<HIPSpace, HIP::scratch_memory_space> {
+  enum : bool { assignable = false };
+  enum : bool { accessible = true };
+  enum : bool { deepcopy = false };
+};
+}  // namespace Impl
+
+namespace Tools {
 namespace Experimental {
-using HIPSpace           = ::Kokkos::HIPSpace;
-using HIPHostPinnedSpace = ::Kokkos::HIPHostPinnedSpace;
-using HIPManagedSpace    = ::Kokkos::HIPManagedSpace;
-using HIP                = ::Kokkos::HIP;
+template <>
+struct DeviceTypeTraits<HIP> {
+  static constexpr DeviceType id = DeviceType::HIP;
+  static int device_id(const HIP& exec) { return exec.hip_device(); }
+};
 }  // namespace Experimental
+}  // namespace Tools
+
+namespace Impl {
+template <class DT, class... DP>
+struct ZeroMemset<HIP, DT, DP...> {
+  ZeroMemset(const HIP& exec_space, const View<DT, DP...>& dst,
+             typename View<DT, DP...>::const_value_type&) {
+    KOKKOS_IMPL_HIP_SAFE_CALL(hipMemsetAsync(
+        dst.data(), 0,
+        dst.size() * sizeof(typename View<DT, DP...>::value_type),
+        exec_space.hip_stream()));
+  }
+
+  ZeroMemset(const View<DT, DP...>& dst,
+             typename View<DT, DP...>::const_value_type&) {
+    KOKKOS_IMPL_HIP_SAFE_CALL(
+        hipMemset(dst.data(), 0,
+                  dst.size() * sizeof(typename View<DT, DP...>::value_type)));
+  }
+};
+}  // namespace Impl
 }  // namespace Kokkos
 
 #endif
